Thu, Jul 4, 2024, 9:33 PM +05:30

+++ world/rough2.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-    float c1 = 12;
+    int c1 = 12;
     char c2 = 'a';
     int c3 = 1333;
